# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ToBu-Finance API (v2 schema)
  version: 1.0.1
  description: |
    REST API for ToBu-Finance using the v2 data model.
    Conventions:
      - ObjectIds are 24-hex strings
      - Money values are integer cents
      - Month inputs are strings in YYYY-MM (server stores month as Date at UTC 1st day)
servers:
  - url: http://localhost:4000
    description: Local dev
paths:
...
  /api/transactions:
    get:
      summary: List transactions
      parameters:
        - in: query
          name: accountId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: month
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: monthFrom
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: monthTo
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: status
          schema:
            type: string
            enum: [booked, pending]
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
  /api/transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete transaction
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/recurrences:
    get:
      summary: List recurrences
      parameters:
        - in: query
          name: accountId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: activeOn
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: nextFrom
          schema: { type: string, format: date-time }
        - in: query
          name: nextTo
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Recurrences
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Recurrence' }
    post:
      summary: Create recurrence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecurrenceCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recurrence' }
  /api/recurrences/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update recurrence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecurrenceUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recurrence' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete recurrence
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: 'NOT_FOUND' }
  schemas:
    ObjectId:
      type: string
      pattern: "^[a-fA-F\\d]{24}$"
      example: "60d21b4667d0d8992e610c85"
    Month:
      type: string
      pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      example: "2025-08"
...
    Transaction:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        accountId: { $ref: '#/components/schemas/ObjectId' }
        categoryId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        title: { type: string, nullable: true }
        type: { type: string, enum: [income, expense] }
        amountCents: { type: integer, minimum: 0 }
        month: { type: string, format: date-time, nullable: true }
        bookDate: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [booked, pending], default: pending }
        isFromSharedAccount: { type: boolean, nullable: true }
        paidByMemberId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        recurrenceId: { $ref: '#/components/schemas/ObjectId', nullable: true }
    TransactionCreate:
      type: object
      required: [accountId, type, amountCents]
      properties:
        accountId: { $ref: '#/components/schemas/ObjectId' }
        categoryId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        title: { type: string, nullable: true }
        type: { type: string, enum: [income, expense] }
        amountCents: { type: integer, minimum: 0 }
        month: { $ref: '#/components/schemas/Month' }
        bookDate: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [booked, pending], default: pending }
        isFromSharedAccount: { type: boolean, nullable: true }
        paidByMemberId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        recurrenceId: { $ref: '#/components/schemas/ObjectId', nullable: true }
    TransactionUpdate:
      allOf: [ { $ref: '#/components/schemas/TransactionCreate' } ]
    RecurrenceSchedule:
      type: object
      properties:
        freq: { type: string, enum: [monthly], default: monthly }
        dayOfMonth: { type: integer, minimum: 1, maximum: 28, default: 1 }
    Recurrence:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        accountId: { $ref: '#/components/schemas/ObjectId' }
        categoryId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        title: { type: string, nullable: true }
        type: { type: string, enum: [income, expense] }
        amountCents: { type: integer, minimum: 0 }
        schedule: { $ref: '#/components/schemas/RecurrenceSchedule' }
        activeFrom: { type: string, format: date-time }
        activeUntil: { type: string, format: date-time, nullable: true }
        createdByMemberId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        nextPlanned: { type: string, format: date-time, nullable: true }
        lastEmitted: { type: string, format: date-time, nullable: true }
    RecurrenceCreate:
      type: object
      required: [accountId, type, amountCents]
      properties:
        accountId: { $ref: '#/components/schemas/ObjectId' }
        categoryId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        title: { type: string, nullable: true }
        type: { type: string, enum: [income, expense] }
        amountCents: { type: integer, minimum: 0 }
        schedule: { $ref: '#/components/schemas/RecurrenceSchedule' }
        activeFrom: { $ref: '#/components/schemas/Month' }
        activeUntil: { $ref: '#/components/schemas/Month' }
        createdByMemberId: { $ref: '#/components/schemas/ObjectId', nullable: true }
        nextPlanned: { type: string, format: date-time, nullable: true }
        lastEmitted: { type: string, format: date-time, nullable: true }
    RecurrenceUpdate:
      allOf: [ { $ref: '#/components/schemas/RecurrenceCreate' } ]
...